moved formatting to utilities
