+++ packages/utilities/src/index.ts
@@ -8,6 +8,7 @@ export {
   removeLineBreaksAtStartAndEnd,
   removeSpecialCharacters,
   singleQuotify,
+  formatText,
 } from "./functions/formatting.js";
 export {
   buildCodemodMetadataHash,
